chore: move api keys to org level

+++ apps/web/app/api/v1/management/responses/[responseId]/route.ts
@@ -1,31 +1,9 @@
-import { authenticateRequest, handleErrorResponse } from "@/app/api/v1/auth";
+import { authenticateRequest, handleErrorResponse, hasPermission } from "@/app/api/v1/auth";
 import { responses } from "@/app/lib/api/response";
 import { transformErrorToDetails } from "@/app/lib/api/validator";
-import { hasUserEnvironmentAccess } from "@formbricks/lib/environment/auth";
 import { deleteResponse, getResponse, updateResponse } from "@formbricks/lib/response/service";
 import { getSurvey } from "@formbricks/lib/survey/service";
-import { TResponse, ZResponseUpdateInput } from "@formbricks/types/responses";
-
-const fetchAndValidateResponse = async (authentication: any, responseId: string): Promise<TResponse> => {
-  const response = await getResponse(responseId);
-  if (!response || !(await canUserAccessResponse(authentication, response))) {
-    throw new Error("Unauthorized");
-  }
-  return response;
-};
-
-const canUserAccessResponse = async (authentication: any, response: TResponse): Promise<boolean> => {
-  const survey = await getSurvey(response.surveyId);
-  if (!survey) return false;
-
-  if (authentication.type === "session") {
-    return await hasUserEnvironmentAccess(authentication.session.user.id, survey.environmentId);
-  } else if (authentication.type === "apiKey") {
-    return survey.environmentId === authentication.environmentId;
-  } else {
-    throw Error("Unknown authentication type");
-  }
-};
+import { ZResponseUpdateInput } from "@formbricks/types/responses";
 
 export const GET = async (
   request: Request,
@@ -35,11 +13,19 @@ export const GET = async (
   try {
     const authentication = await authenticateRequest(request);
     if (!authentication) return responses.notAuthenticatedResponse();
-    const response = await fetchAndValidateResponse(authentication, params.responseId);
-    if (response) {
-      return responses.successResponse(response);
+    const response = await getResponse(params.responseId);
+    if (!response) {
+      return responses.notFoundResponse("Response", params.responseId);
     }
-    return responses.notFoundResponse("Response", params.responseId);
+    const survey = await getSurvey(response.surveyId);
+    if (!survey) {
+      return responses.notFoundResponse("Survey", response.surveyId, true);
+    }
+
+    if (!hasPermission(authentication.environmentPermissions, survey.environmentId, "GET")) {
+      return responses.unauthorizedResponse();
+    }
+    return responses.successResponse(response);
   } catch (error) {
     return handleErrorResponse(error);
   }
@@ -53,9 +39,16 @@ export const DELETE = async (
   try {
     const authentication = await authenticateRequest(request);
     if (!authentication) return responses.notAuthenticatedResponse();
-    const response = await fetchAndValidateResponse(authentication, params.responseId);
+    const response = await getResponse(params.responseId);
     if (!response) {
       return responses.notFoundResponse("Response", params.responseId);
+    }
+    const survey = await getSurvey(response.surveyId);
+    if (!survey) {
+      return responses.notFoundResponse("Survey", response.surveyId, true);
+    }
+    if (!hasPermission(authentication.environmentPermissions, survey.environmentId, "DELETE")) {
+      return responses.unauthorizedResponse();
     }
     const deletedResponse = await deleteResponse(params.responseId);
     return responses.successResponse(deletedResponse);
@@ -72,7 +65,17 @@ export const PUT = async (
   try {
     const authentication = await authenticateRequest(request);
     if (!authentication) return responses.notAuthenticatedResponse();
-    await fetchAndValidateResponse(authentication, params.responseId);
+    const response = await getResponse(params.responseId);
+    if (!response) {
+      return responses.notFoundResponse("Response", params.responseId);
+    }
+    const survey = await getSurvey(response.surveyId);
+    if (!survey) {
+      return responses.notFoundResponse("Survey", response.surveyId, true);
+    }
+    if (!hasPermission(authentication.environmentPermissions, survey.environmentId, "PUT")) {
+      return responses.unauthorizedResponse();
+    }
     let responseUpdate;
     try {
       responseUpdate = await request.json();
